Break up AppNode into separate files
